requirements.txt
boto3>=1.26.0
pandas>=1.5.0
pyarrow>=12.0.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0 